requirements.txt
langchain>=0.3.7
langchain-huggingface>=0.3.1
langchain-anthropic>=0.2.4
langchain-community>=0.3.5
langgraph>=0.2.45
pydantic>=2.10.3
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
pytest>=8.3.4
pytest-cov>=6.0.0
gherkin-official>=28.0.0
loguru>=0.7.3
playwright>=1.48.0
